pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "privacy-metrics"
version = "0.1.0"
description = "Privacy metrics utilities"
authors = [{ name = "Aindo Spa", email = "hello@aindo.com" }]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "anonymeter",
    "anyio",
    "appnope",
    "argon2-cffi",
    "argon2-cffi-bindings",
    "arrow",
    "asttokens",
    "async-lru",
    "attrs",
    "babel",
    "beautifulsoup4",
    "bleach",
    "certifi",
    "cffi",
    "charset-normalizer",
    "comm",
    "contourpy",
    "cycler",
    "debugpy",
    "decorator",
    "defusedxml",
    "exceptiongroup",
    "executing",
    "faiss-cpu",
    "fastjsonschema",
    "filelock",
    "fonttools",
    "fqdn",
    "fsspec",
    "h11",
    "httpcore",
    "httpx",
    "idna",
    "ipykernel",
    "ipython",
    "isoduration",
    "jedi",
    "Jinja2",
    "joblib",
    "json5",
    "jsonpointer",
    "jsonschema",
    "jsonschema-specifications",
    "jupyter-events",
    "jupyter-lsp",
    "jupyter_client",
    "jupyter_core",
    "jupyter_server",
    "jupyter_server_terminals",
    "jupyterlab",
    "jupyterlab_pygments",
    "jupyterlab_server",
    "kiwisolver",
    "llvmlite",
    "MarkupSafe",
    "matplotlib",
    "matplotlib-inline",
    "mistune",
    "mpmath",
    "nbclient",
    "nbconvert",
    "nbformat",
    "nest-asyncio",
    "networkx",
    "notebook",
    "notebook_shim",
    "numba",
    "numpy",
    "overrides",
    "packaging",
    "pandas",
    "pandocfilters",
    "parso",
    "pexpect",
    "pillow",
    "platformdirs",
    "privacy_eval @ git+https://github.com/aindo-com/privacy-eval-paper.git@6d054c482349a92f41c2807bbcf214d2b66fb006",
    "prometheus_client",
    "prompt_toolkit",
    "psutil",
    "ptyprocess",
    "pure_eval",
    "pycparser",
    "Pygments",
    "pyparsing",
    "python-dateutil",
    "python-json-logger",
    "pytz",
    "PyYAML",
    "pyzmq",
    "referencing",
    "requests",
    "rfc3339-validator",
    "rfc3986-validator",
    "rpds-py",
    "scikit-learn",
    "scipy",
    "Send2Trash",
    "six",
    "sniffio",
    "soupsieve",
    "stack-data",
    "sympy",
    "terminado",
    "threadpoolctl",
    "tinycss2",
    "tomli",
    "torch",
    "tornado",
    "tqdm",
    "traitlets",
    "types-python-dateutil",
    "typing_extensions",
    "tzdata",
    "uri-template",
    "urllib3",
    "wcwidth",
    "webcolors",
    "webencodings",
    "websocket-client",
    "xgboost",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["privacy_metrics"]